Refactor ellipsoid.js, putting ray intersections in raygun.js
